requirements.txt
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.2
